+++ kaithem/api/chandler.py
@@ -1,9 +1,8 @@
 from typing import Any, Callable
 
-from scullery import workers
-
 from kaithem.src.chandler import global_actions as _global_actions
 from kaithem.src.chandler import groups as _groups
+from kaithem.src.chandler.core import serialized_async_with_core_lock
 
 
 def add_command(name: str, f: Callable):
@@ -20,7 +19,7 @@ def trigger_event(event: str, value: Any = None):
     def f():
         _global_actions.cl_event(event, value)
 
-    workers.do(f)
+    serialized_async_with_core_lock(f)
 
 
 def shortcut(s: str):
@@ -29,4 +28,4 @@ def shortcut(s: str):
     def f():
         _global_actions.cl_trigger_shortcut_code(s)
 
-    workers.do(f)
+    serialized_async_with_core_lock(f)